+++ commonSrc/net/xuset/tSquare/imaging/TransformedGraphics.java
@@ -55,7 +55,7 @@ public class TransformedGraphics implements IGraphics{
 	}
 
 	@Override
-	public void drawRoundedRect(float x, float y, float w, float h, int rx, int ry) {
+	public void drawRoundedRect(float x, float y, float w, float h, float rx, float ry) {
 		g.drawRoundedRect(
 				transformX(x),
 				transformY(y),
@@ -89,7 +89,7 @@ public class TransformedGraphics implements IGraphics{
 	}
 
 	@Override
-	public void fillRoundedRect(float x, float y, float w, float h, int rx, int ry) {
+	public void fillRoundedRect(float x, float y, float w, float h, float rx, float ry) {
 		g.fillRoundedRect(
 				transformX(x),
 				transformY(y),
